Introduce resizable Panel
This major refactor puts all the form components and text parts of the
app inside a new Panel component.

Next steps:
* Add button to toggle hide/show state for Panel
* Possibly separate Object List from "content" area, each in their own
  boxes
