+++ mathplayground/templates/index.html
@@ -15,7 +15,6 @@
               integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi"
               crossorigin="anonymous">
 
-
         <link rel="stylesheet" href="{% static 'mathplayground/global.css' %}">
         <link rel="stylesheet" href="{% static 'mathplayground/build/bundle.css' %}">
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
