+++ media/src/objects/Point.svelte
@@ -48,7 +48,6 @@
     export let show = true;
 
     export let scene;
-    export let shadeUp;
     export let render = () => {};
     export let onClose = () => {};
     export let gridStep;
@@ -116,13 +115,11 @@
     });
 
     const shiftDown = (e) => {
-        if (shadeUp) {
-            switch (e.key) {
-                case 'Backspace':
-                    point.visible = !point.visible;
-                    render();
-                    break;
-            }
+        switch (e.key) {
+            case 'Backspace':
+                point.visible = !point.visible;
+                render();
+                break;
         }
     };
